Organize CMake project
