+++ CMakeLists.txt
@@ -1,19 +1,34 @@
-cmake_minimum_required(VERSION 3.1)
-
+cmake_minimum_required(VERSION 3.16)
 project(SFML_GUI)
 
-#file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/assets DESTINATION ${CMAKE_CURRENT_BINARY_DIR}/src)
-find_package(SFML 2.5 COMPONENTS graphics window system REQUIRED)
+set(CMAKE_CXX_STANDARD 17)
 
-include_directories(include/headers)
+option(BUILD_SHARED_LIBS "Build shared libraries" OFF)
 
-set(SOURCES
+set(INCLUDE_DIR ${PROJECT_SOURCE_DIR}/include)
+set(BIN_DIR ${PROJECT_SOURCE_DIR}/bin)
+set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${BIN_DIR})
+set(EXECUTABLE_OUTPUT_PATH ${BIN_DIR})
+
+add_subdirectory(vendor)
+include_directories(${INCLUDE_DIR})
+
+file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/assets DESTINATION ${BIN_DIR})
+
+set(SOURCE_FILES
     src/main.cpp
-    src/Button.cpp
+    src/ui/Button.cpp
 )
+add_executable(${PROJECT_NAME} ${SOURCE_FILES})
+target_link_libraries(${PROJECT_NAME} PRIVATE sfml-graphics sfml-window sfml-system)
 
-set(EXECUTABLE_OUTPUT_PATH ../bin)
+# Necessary for SFML
+if(WIN32)
+    add_custom_command(
+        TARGET ${PROJECT_NAME}
+        COMMENT "Copy OpenAL DLL"
+        PRE_BUILD COMMAND ${CMAKE_COMMAND} -E copy ${SFML_SOURCE_DIR}/extlibs/bin/$<IF:$<EQUAL:${CMAKE_SIZEOF_VOID_P},8>,x64,x86>/openal32.dll $<TARGET_FILE_DIR:${PROJECT_NAME}>
+        VERBATIM)
+endif()
 
-add_executable(SFML_GUI ${SOURCES})
-
-target_link_libraries(SFML_GUI sfml-graphics sfml-window sfml-system)
+install(TARGETS ${PROJECT_NAME})
